integrate master and resolve conflicts
