+++ .github/workflows/npm-build.yml
@@ -16,7 +16,7 @@ jobs:
       - name: Use Node.js
         uses: actions/setup-node@v4
         with:
-          node-version: "20.x"
+          node-version: "22.x"
       - run: npm ci
       - run: npm run build
       - name: Cypress E2E
