+++ package.json
@@ -5,9 +5,9 @@
   "private": true,
   "dependencies": {
     "@auth0/auth0-spa-js": "2.1.3",
-    "@azure/msal-browser": "3.26.1",
+    "@azure/msal-browser": "3.27.0",
     "@chakra-ui/icons": "2.2.4",
-    "@chakra-ui/react": "2.10.3",
+    "@chakra-ui/react": "2.10.4",
     "@chakra-ui/theme-tools": "2.2.6",
     "@emotion/react": "11.13.3",
     "@emotion/styled": "11.13.0",
@@ -20,8 +20,8 @@
     "dayzed": "3.2.3",
     "final-form": "4.20.10",
     "final-form-arrays": "3.1.0",
-    "framer-motion": "11.11.9",
-    "i18next": "23.16.3",
+    "framer-motion": "11.11.11",
+    "i18next": "23.16.5",
     "js-cookie": "3.0.5",
     "lodash-es": "4.17.21",
     "ramda": "0.30.1",
@@ -34,11 +34,11 @@
     "react-final-form": "6.5.9",
     "react-final-form-arrays": "3.1.4",
     "react-helmet-async": "2.0.5",
-    "react-i18next": "15.1.0",
+    "react-i18next": "15.1.1",
     "react-icons": "5.3.0",
     "react-if": "4.1.5",
     "react-redux": "9.1.2",
-    "react-router-dom": "6.27.0",
+    "react-router-dom": "6.28.0",
     "usehooks-ts": "3.1.0",
     "zod": "3.23.8",
     "zod-i18n-map": "2.27.0"
@@ -65,60 +65,65 @@
     "e2e:ci:run": "cross-env CYPRESS_BASE_URL=http://localhost:4321 npm run cypress:run"
   },
   "devDependencies": {
-    "@cypress/code-coverage": "3.13.4",
+    "@cypress/code-coverage": "3.13.6",
+    "@eslint/js": "^8.57.1",
     "@iodigital/vite-plugin-msw": "^2.0.0",
     "@svgr/plugin-jsx": "8.1.0",
     "@svgr/plugin-svgo": "8.1.0",
-    "@swc/plugin-emotion": "4.0.3",
+    "@swc/plugin-emotion": "7.0.0",
     "@testing-library/cypress": "10.0.2",
-    "@testing-library/jest-dom": "6.6.2",
+    "@testing-library/jest-dom": "6.6.3",
     "@testing-library/react": "16.0.1",
     "@testing-library/user-event": "14.5.2",
+    "@types/eslint__js": "^8.42.3",
     "@types/jest": "29.5.14",
     "@types/js-cookie": "3.0.6",
     "@types/lodash-es": "4.17.12",
-    "@types/node": "22.7.9",
+    "@types/node": "22.9.0",
     "@types/ramda": "0.30.2",
     "@types/react": "18.3.12",
     "@types/react-copy-to-clipboard": "5.0.7",
     "@types/react-dom": "18.3.1",
-    "@typescript-eslint/eslint-plugin": "8.11.0",
-    "@typescript-eslint/parser": "8.11.0",
+    "@typescript-eslint/eslint-plugin": "8.13.0",
+    "@typescript-eslint/parser": "8.13.0",
     "@vite-pwa/assets-generator": "0.2.6",
     "@vitejs/plugin-react-swc": "3.7.1",
-    "@vitest/coverage-v8": "2.1.3",
-    "@vitest/ui": "2.1.3",
-    "concurrently": "9.0.1",
+    "@vitest/coverage-v8": "2.1.4",
+    "@vitest/ui": "2.1.4",
+    "concurrently": "9.1.0",
     "cross-env": "7.0.3",
-    "cypress": "13.15.0",
+    "cypress": "13.15.2",
     "cypress-localstorage-commands": "2.2.6",
     "cypress-wait-until": "3.0.2",
     "dotenv": "16.4.5",
+    "eslint": "^8.57.1",
     "eslint-import-resolver-typescript": "3.6.3",
-    "eslint-plugin-cypress": "^3.5.0",
+    "eslint-plugin-chai-friendly": "^1.0.1",
+    "eslint-plugin-cypress": "^3.6.0",
     "eslint-plugin-import": "2.31.0",
-    "eslint-plugin-jsx-a11y": "6.10.1",
+    "eslint-plugin-jsx-a11y": "6.10.2",
+    "eslint-plugin-mocha": "^10.5.0",
     "eslint-plugin-prettier": "5.2.1",
     "eslint-plugin-react": "7.37.2",
     "eslint-plugin-react-hooks": "5.0.0",
-    "eslint-plugin-react-refresh": "0.4.13",
-    "msw": "2.4.9",
+    "eslint-plugin-react-refresh": "0.4.14",
+    "msw": "2.6.4",
     "prettier": "3.3.3",
     "rollup-plugin-copy": "3.5.0",
     "sharp": "0.33.5",
     "svgo": "3.3.2",
     "typescript": "5.6.3",
-    "vite": "5.4.10",
+    "vite": "5.4.11",
     "vite-plugin-eslint": "1.8.1",
-    "vite-plugin-i18n-ally": "5.1.3",
+    "vite-plugin-i18n-ally": "5.2.2",
     "vite-plugin-image-optimizer": "1.1.8",
     "vite-plugin-legacy-swc": "1.2.1",
     "vite-plugin-preload": "0.4.2",
     "vite-plugin-pwa": "0.20.5",
     "vite-plugin-react-click-to-component": "3.0.0",
-    "vite-plugin-static-copy": "2.0.0",
-    "vite-plugin-svgr": "4.2.0",
-    "vitest": "2.1.3",
+    "vite-plugin-static-copy": "2.1.0",
+    "vite-plugin-svgr": "4.3.0",
+    "vitest": "2.1.4",
     "wait-on": "8.0.1",
     "web-vitals": "4.2.4"
   },
@@ -129,7 +134,7 @@
   },
   "engineStrict": true,
   "engines": {
-    "node": "20.x",
+    "node": "22.x",
     "npm": ">=10"
   },
   "msw": {
